mostly finished with demos, moved some methods around in files
