+++ PARCtorch/data/dataset.py
@@ -314,100 +314,7 @@ def custom_collate_fn(batch):
 
     return ic, t0, t1, target
 
-def visualize_channels(ic, t0, t1, target, channel_names=None, channel_cmaps=None, sample_index=0, batch_size=1, figsize=(25, 20)):
-    """
-    Visualizes the channels of the initial condition and target timesteps with customizable color maps.
-    
-    Args:
-        ic (torch.Tensor): Initial condition tensor of shape (batch_size, channels, height, width).
-        t0 (torch.Tensor): Scalar tensor (0.0), not used in visualization but kept for compatibility.
-        t1 (torch.Tensor): Tensor of shape (future_steps,), not used directly in visualization.
-        target (torch.Tensor): Target tensor of shape (future_steps, batch_size, channels, height, width).
-        channel_names (list of str, optional): List of channel names for labeling. If None, channels will be unnamed.
-        channel_cmaps (list of str or matplotlib.colors.Colormap, optional): 
-            List of color maps for each channel. If None, 'viridis' is used for all channels.
-        sample_index (int, optional): Index of the sample in the batch to visualize. Defaults to 0.
-        batch_size (int, optional): Total number of samples in the batch. Needed if `sample_index` is to be visualized.
-        figsize (tuple, optional): Figure size for the plots. Defaults to (25, 20).
-    
-    Raises:
-        ValueError: If `sample_index` is out of bounds or if `channel_cmaps` length doesn't match number of channels.
-    """
-    # Validate sample_index
-    if sample_index >= batch_size or sample_index < 0:
-        raise ValueError(f"sample_index {sample_index} is out of bounds for batch size {batch_size}.")
-    
-    # Select the specific sample from the batch
-    ic_sample = ic[sample_index]             # Shape: (channels, height, width)
-    target_sample = target[:, sample_index]   # Shape: (future_steps, channels, height, width)
-    
-    num_channels = ic_sample.shape[0]
-    future_steps = target_sample.shape[0]
-    
-    # Handle channel_cmaps
-    if channel_cmaps is None:
-        # Use 'viridis' for all channels by default
-        channel_cmaps = ['viridis'] * num_channels
-    else:
-        if not isinstance(channel_cmaps, list):
-            raise TypeError("channel_cmaps must be a list of color map names or Colormap objects.")
-        if len(channel_cmaps) != num_channels:
-            raise ValueError(
-                f"Length of channel_cmaps ({len(channel_cmaps)}) does not match number of channels ({num_channels}). "
-                f"Ensure you provide a color map for each channel or leave it as None to use default."
-            )
-    
-    # Debugging: Print min and max of each channel
-    print("Channel Data Statistics:")
-    for idx in range(num_channels):
-        channel_data_ic = ic_sample[idx].cpu().numpy()
-        channel_data_target = target_sample[:, idx].cpu().numpy()
-        print(f"Channel {idx}: IC min={channel_data_ic.min()}, IC max={channel_data_ic.max()}")
-        for step in range(future_steps):
-            print(f"  Step {step + 1}: min={channel_data_target[step].min()}, max={channel_data_target[step].max()}")
-    
-    # Determine subplot grid size
-    cols = num_channels
-    rows = future_steps + 1  # +1 for the initial condition
-    
-    # Create a figure
-    fig, axes = plt.subplots(rows, cols, figsize=figsize)
-    
-    # Handle axes array shape
-    if num_channels == 1:
-        axes = axes.reshape(-1, 1)
-    if rows == 1:
-        axes = axes.reshape(1, -1)
-    
-    # Plot Initial Condition
-    for channel in range(num_channels):
-        ax = axes[0, channel]
-        data = ic_sample[channel].cpu().numpy()
-        cmap = channel_cmaps[channel]
-        im = ax.imshow(data, cmap=cmap)
-        if channel_names:
-            ax.set_title(f"IC - {channel_names[channel]}")
-        else:
-            ax.set_title(f"IC - Channel {channel}")
-        ax.axis('off')
-        fig.colorbar(im, ax=ax, fraction=0.046, pad=0.04)
-    
-    # Plot Targets
-    for step in range(future_steps):
-        for channel in range(num_channels):
-            ax = axes[step + 1, channel]
-            data = target_sample[step, channel].cpu().numpy()
-            cmap = channel_cmaps[channel]
-            im = ax.imshow(data, cmap=cmap)
-            if channel_names:
-                ax.set_title(f"Step {step + 1} - {channel_names[channel]}")
-            else:
-                ax.set_title(f"Step {step + 1} - Channel {channel}")
-            ax.axis('off')
-            fig.colorbar(im, ax=ax, fraction=0.046, pad=0.04)
-    
-    plt.tight_layout()
-    plt.show()
+
     
     
     
@@ -418,17 +325,18 @@ class InitialConditionDataset(Dataset):
     Each sample consists of:
         - ic: Initial condition tensor of shape (channels, height, width)
         - t0: Scalar tensor (0.0)
-        - t1: Tensor indicating future steps (e.g., [1/future_steps, 2/future_steps, ..., future_steps/future_steps])
+        - t1: Tensor indicating future steps (defined by the user or calculated automatically)
         - target: Placeholder or can be set to None since the model will predict the entire sequence
     """
 
-    def __init__(self, data_dirs, future_steps=1, min_max_path=None, required_channels=None, validate=True):
+    def __init__(self, data_dirs, future_steps=1, t1=None, min_max_path=None, required_channels=None, validate=True):
         """
         Initializes the InitialConditionDataset.
 
         Args:
             data_dirs (list of str): List of directories containing preprocessed `.npy` files.
             future_steps (int): Number of timesteps in the future the model will predict.
+            t1 (list, optional): A list of future steps for prediction. If None, evenly spaced steps will be calculated.
             min_max_path (str, optional): Path to the JSON file containing min and max values for each channel.
                                           If None, it will look for 'min_max.json' in each data directory.
             required_channels (int, optional): Number of channels expected in the data.
@@ -522,11 +430,17 @@ class InitialConditionDataset(Dataset):
                 )
             self.timesteps_per_file.append(timesteps)
 
-        # Precompute t1
-        self.t1 = torch.tensor(
-            [(i + 1) / self.future_steps for i in range(self.future_steps)],
-            dtype=torch.float32
-        )  # Shape: (future_steps,)
+        # Handle t1 logic: either use the user-defined list or generate evenly spaced steps
+        if t1 is None:
+            self.t1 = torch.tensor(
+                [(i + 1) / self.future_steps for i in range(self.future_steps)],
+                dtype=torch.float32
+            )  # Shape: (future_steps,)
+        else:
+            if len(t1) != future_steps:
+                raise ValueError(f"Length of 't1' ({len(t1)}) must match 'future_steps' ({future_steps}).")
+            self.t1 = torch.tensor(t1, dtype=torch.float32)  # User-defined t1
+
         self.t0 = torch.tensor(0.0, dtype=torch.float32)  # Scalar
 
         # Initialize a cache for memory-mapped files to improve performance
@@ -598,12 +512,6 @@ class InitialConditionDataset(Dataset):
         ic = data_tensor  # Shape: (channels, height, width)
 
         return ic, self.t0, self.t1, None  # Target is None since the model will predict it
-
-    def __del__(self):
-        # Close all memmap files when the dataset is deleted
-        for memmap in self._memmap_cache.values():
-            del memmap
-        self._memmap_cache.clear()
         
 def initial_condition_collate_fn(batch):
     """
